+++ test/securx_method_channel_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/services.dart';
-import 'package:securx/app_guard_method_channel.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:securx/securx_method_channel.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();